requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
colorama==0.4.4
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
importlib-metadata==3.4.0
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.6.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
wrapt==1.12.1
zipp==3.4.0
